# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-aliyunmc",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/task": {
            "get": {
                "description": "根据指定的任务类型，获取数据库中的一条正在进行的相应类型任务记录，可以选择性地包含与该任务记录关联的推送事件内容。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "获取指定类型的一条运行任务记录",
                "parameters": [
                    {
                        "description": "获取任务记录请求体",
                        "name": "gettaskquery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.GetTaskQuery"
                        }
                    },
                    {
                        "type": "string",
                        "description": "任务类型",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-tasks_GetTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
        "/task/cancel/{taskId}": {
            "get": {
                "description": "根据任务标识符，尝试取消一个可能正在运行的任务",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks",
                    "admin"
                ],
                "summary": "取消任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要取消的任务标识符",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-bool"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
        "/task/overview": {
            "get": {
                "description": "返回系统任务模块的总览信息，如总任务数量、成功执行数量等。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "获取任务总览",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-tasks_TaskOverview"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
        "/task/s": {
            "get": {
                "description": "根据传入的查询参数，返回多条任务记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "获取多条任务记录",
                "parameters": [
                    {
                        "description": "分页参数等",
                        "name": "gettasksquery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.GetTasksQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-array_store_JoinedTask"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
        "/task/{taskId}": {
            "get": {
                "description": "根据标识符查询数据库中一条任务记录，可以选择性地包含与该任务记录关联的推送事件内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "获取一条任务记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标任务标识符",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "获取任务记录请求体",
                        "name": "gettaskquery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.GetTaskQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-tasks_GetTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "根据传入的凭证返回该用户的非敏感信息，例如用户名、ID、创建时间、权限组等信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "获取当前已登入用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.DataResp-store_User"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            },
            "post": {
                "description": "根据提供的用户名和密码创建一个新的用户。用户名是唯一的，如果与现有用户重复则返回409。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "创建用户请求体",
                        "name": "createuserrequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "consts.TaskStatus": {
            "type": "string",
            "enum": [
                "running",
                "success",
                "failed",
                "cancelled",
                "timed_out"
            ],
            "x-enum-varnames": [
                "TaskStatusRunning",
                "TaskStatusSuccess",
                "TaskStatusFailed",
                "TaskStatusCancelled",
                "TaskStatusTimedOut"
            ]
        },
        "consts.TaskType": {
            "type": "string",
            "enum": [
                "instance_deployment"
            ],
            "x-enum-varnames": [
                "TaskTypeInstanceDeployment"
            ]
        },
        "events.Event": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content 是该事件的内容，可能是序列化以后的字符串",
                    "type": "string"
                },
                "createdAt": {
                    "description": "CreatedAt 是事件的创建时间（由数据库记录）",
                    "type": "string"
                },
                "isError": {
                    "description": "IsError 决定前端该事件的显示效果。IsError 并不一定表示此事件与系统的严重错误有关。",
                    "type": "boolean"
                },
                "isPublic": {
                    "description": "IsPublic 决定该事件是否可以发送给未登录用户",
                    "type": "boolean"
                },
                "ord": {
                    "description": "Ord 是该事件在该任务内部的顺序（按照时间排序）。顺序的开始是任意的，但不可重复。",
                    "type": "integer"
                },
                "taskId": {
                    "description": "TaskId 是与该事件相关联的任务标识符，并充当 Ord 的命名空间",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/events.EventType"
                }
            }
        },
        "events.EventType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "TypeDeployment",
                "TypeServer",
                "TypeInstance",
                "TypeError",
                "TypeSync"
            ]
        },
        "helpers.DataResp-array_store_JoinedTask": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 为该结构中的数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.JoinedTask"
                    }
                }
            }
        },
        "helpers.DataResp-bool": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 为该结构中的数据",
                    "type": "boolean"
                }
            }
        },
        "helpers.DataResp-store_User": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 为该结构中的数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/store.User"
                        }
                    ]
                }
            }
        },
        "helpers.DataResp-tasks_GetTaskResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 为该结构中的数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/tasks.GetTaskResponse"
                        }
                    ]
                }
            }
        },
        "helpers.DataResp-tasks_TaskOverview": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 为该结构中的数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/tasks.TaskOverview"
                        }
                    ]
                }
            }
        },
        "helpers.ErrorResp": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details 为错误信息",
                    "type": "string"
                }
            }
        },
        "store.JoinedTask": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/consts.TaskStatus"
                },
                "type": {
                    "$ref": "#/definitions/consts.TaskType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "store.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "tasks.GetTaskQuery": {
            "type": "object",
            "properties": {
                "withJoinedPushedEvents": {
                    "description": "WithJoinedPushedEvents 指定是否在结果中包含与该任务相关的所有推送事件内容的合并值",
                    "type": "boolean"
                },
                "withPushedEvents": {
                    "description": "WithPushedEvents 指定是否在结果中包含与该任务相关的所有推送事件",
                    "type": "boolean"
                }
            }
        },
        "tasks.GetTaskResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "joinedPushedEvents": {
                    "description": "JoinedPushedEvents 由 GetTaskQuery.WithJoinedPushedEvents 指定是否包含",
                    "type": "string"
                },
                "pushedEvents": {
                    "description": "PushedEvents 由 GetTaskQuery.WithPushedEvents 指定是否包含",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/events.Event"
                    }
                },
                "status": {
                    "$ref": "#/definitions/consts.TaskStatus"
                },
                "type": {
                    "$ref": "#/definitions/consts.TaskType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "tasks.GetTasksQuery": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "pageSize": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "tasks.TaskOverview": {
            "type": "object",
            "properties": {
                "latest": {
                    "description": "Latest 是系统中最新一条任务的信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/store.JoinedTask"
                        }
                    ]
                },
                "successCount": {
                    "description": "SuccessCount 是成功任务的总数量",
                    "type": "integer"
                },
                "unsuccessCount": {
                    "description": "UnsuccessCount 是失败任务的总数量",
                    "type": "integer"
                }
            }
        },
        "users.CreateUserRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password 是要创建用户的密码明文",
                    "type": "string"
                },
                "username": {
                    "description": "Username 是要创建用户的用户名，不可与其他用户重复",
                    "type": "string"
                }
            }
        },
...
    }
}